Move code to get CLI matches to new function
Moves the code used to get CLI matches from main.rs to a new function in
cli.rs.
